db/tests/requirements.txt
supabase>=2.0.0
pytest>=7.0.0
python-dotenv>=1.0.0
